fix another typo

+++ iris-mpc-gpu/tests/bucket_2d_aggregation.rs
@@ -175,7 +175,7 @@ mod bucket_2d_aggregation_test {
                 let left_reduced = left
                     .into_iter()
                     .reduce(|(c1, m1), (c2, m2)| {
-                        if (c1 as i16 as i64 * m2 as i64) > (c2 as i64 * m1 as i64) {
+                        if (c1 as i16 as i64 * m2 as i64) > (c2 as i16 as i64 * m1 as i64) {
                             (c1, m1)
                         } else {
                             (c2, m2)
